+++ TablePro/Views/Results/DataGridUpdateSnapshot.swift
@@ -20,5 +20,6 @@ struct DataGridUpdateSnapshot: Equatable {
     let rowHeight: CGFloat
     let alternatingRows: Bool
     let reloadVersion: Int
+    let contentRevision: Int
     let showObjectComments: Bool
 }